Add dedicated Temas views for alumno and docente
