+++ frontend/src/app/features/alumno/alumno.routes.ts
@@ -15,6 +15,7 @@ import { AlumnoEvaluationsComponent } from './evaluations/alumno-evaluations.com
 import { AlumnoEntregaComponent } from './entrega/alumno-entrega.component';
 import { AlumnoPracticaComponent } from './practica/alumno-practica.component';
 import { AlumnoViewerComponent } from './viewer/alumno-viewer.component';
+import { AlumnoTemasComponent } from './temas/alumno-temas.component';
 
 export const ALUMNO_ROUTES: Routes = [
   {
@@ -29,6 +30,7 @@ export const ALUMNO_ROUTES: Routes = [
       { path: 'calendar',       component: AlumnoCalendarComponent,      title: 'Alumno | Calendar' },
       { path: 'notifications',  component: AlumnoNotificationsComponent, title: 'Alumno | Notifications' },
       { path: 'trabajo',        component: AlumnoTrabajoComponent,       title: 'Alumno | Trabajo' },
+      { path: 'temas',          component: AlumnoTemasComponent,         title: 'Alumno | Temas' },
       { path: 'perfil',         component: AlumnoPerfilComponent,        title: 'Alumno | Perfil' },
       { path: 'reuniones',      component: AlumnoReunionesComponent,     title: 'Alumno | Reuniones' },
       { path: 'bandeja',        component: AlumnoBandejaComponent,       title: 'Alumno | Bandeja' },
