+++ frontend/src/app/features/docente/calendario/calendario.component.css
@@ -33,7 +33,7 @@
 .doc-title{
   text-align:center; font-weight:900; color:var(--doc-text); font-size:22px;
 }
-.doc-pill{
+.doc-pill{ 
   border:0; background:#e9f1ff; color:var(--doc-text);
   padding:8px 14px; font-weight:800; border-radius:20px; cursor:pointer;
   box-shadow:inset 0 0 0 1px #cfe3ff;
@@ -41,6 +41,15 @@
 .doc-pill:hover{ background:#deebff; }
 .doc-pill.muted{ background:#eef2f7; box-shadow:inset 0 0 0 1px #d9e2ef; }
 .doc-pill.primary{ background:var(--doc-primary); color:#fff; box-shadow:0 6px 16px rgba(25,118,210,.25); }
+.doc-pill[disabled]{
+  opacity:.6;
+  cursor:not-allowed;
+  box-shadow:inset 0 0 0 1px #d1d5db;
+}
+.doc-pill.primary[disabled]{
+  background:linear-gradient(135deg, #a7c8f3, #7ea8e1);
+  color:#f1f5fb;
+}
 .doc-divider{ height:2px; background:#d6e5ff; margin:8px 0 10px; border-radius:2px; }
 
 /* ===== Calendario grande (7 columnas) ===== */
