+++ frontend/src/app/features/docente/layout/docente-layout.component.ts
@@ -27,6 +27,7 @@ export class DocenteLayoutComponent implements OnDestroy {
     { route: 'bandeja', label: 'Bandeja de Entrada', icon: 'assets/Bandeja_entrada.png', alt: 'Bandeja de entrada' },
     { route: 'dashboard', label: 'Panel Docente', icon: 'assets/Inicio.png', alt: 'Panel docente' },
     { route: 'trabajo', label: 'Trabajo de Título', icon: 'assets/Procesos.png', alt: 'Trabajo de título' },
+    { route: 'temas', label: 'Temas', icon: 'assets/Reportes.png', alt: 'Temas de trabajo de título' },
     { route: 'evaluaciones', label: 'Evaluaciones', icon: 'assets/Evaluaciones.png', alt: 'Evaluaciones' },
     { route: 'calendario', label: 'Calendario', icon: 'assets/Calendario.png', alt: 'Calendario' },
     { route: 'perfil', label: 'Perfil', icon: 'assets/Perfil.png', alt: 'Perfil' },
